+++ routes/api/listingstory-routes.js
@@ -1,13 +1,12 @@
 const express = require('express');
 const router = express.Router({ mergeParams: true });
 const { check, oneOf } = require('express-validator');
-const advancedResults = require('../../middleware/advancedResults');
-const { protect } = require('../../middleware/auth');
-const { checkInputError } = require('../../middleware/inputValidation');
 const {
-  NO_FIELD_UPDATED_MSG,
-  INVALID_FIELD_MSG,
-} = require('../../utils/inputExceptionMsg');
+  advancedResults,
+  checkInputError,
+  protect,
+} = require('../../middleware');
+const { NO_FIELD_UPDATED_MSG, INVALID_FIELD_MSG } = require('../../utils');
 
 // import controllers here
 const {