+++ routes/api/profile-routes.js
@@ -1,17 +1,20 @@
 const express = require('express');
 const router = express.Router({ mergeParams: true });
 const { check, oneOf } = require('express-validator');
-
-const advancedResults = require('../../middleware/advancedResults');
-const { protect, authorise } = require('../../middleware/auth');
-const { checkInputError } = require('../../middleware/inputValidation');
-const { mapSingleFileLocation } = require('../../middleware/fileUploadHelper');
 const {
+  advancedResults,
+  checkInputError,
+  protect,
+  authorise,
+  mapSingleFileLocation,
+} = require('../../middleware');
+const {
+  DATETIME_REGEX,
   INVALID_TIMESTAMP_MSG,
   INVALID_FIELD_MSG,
-} = require('../../utils/inputExceptionMsg');
-const { DATETIME_REGEX } = require('../../utils/regex');
-const { uploadFile } = require('../../utils/fileUploader');
+  NO_FIELD_UPDATED_MSG,
+  uploadFile,
+} = require('../../utils');
 
 // import controllers here
 const {
@@ -22,7 +25,6 @@ const {
   verifyProfile,
   uploadPic,
 } = require('../../controllers/profiles');
-const { NO_FIELD_UPDATED_MSG } = require('../../utils/inputExceptionMsg');
 
 // map routes to controller
 router