+++ routes/api/skill-routes.js
@@ -1,14 +1,17 @@
 const express = require('express');
 const router = express.Router({ mergeParams: true });
 const { check } = require('express-validator');
-const advancedResults = require('../../middleware/advancedResults');
-const { protect, authorise } = require('../../middleware/auth');
-const { checkInputError } = require('../../middleware/inputValidation');
-const { ALPHA_WHITESPACE_REGEX } = require('../../utils/regex');
 const {
+  advancedResults,
+  protect,
+  authorise,
+  checkInputError,
+} = require('../../middleware');
+const {
+  ALPHA_WHITESPACE_REGEX,
   INVALID_ALPHA_SPACE_MSG,
   NO_FIELD_UPDATED_MSG,
-} = require('../../utils/inputExceptionMsg');
+} = require('../../utils');
 
 // import controllers here
 const {