+++ routes/test/upload-routes.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
 
-const { mapSingleFileLocation } = require('../../middleware/fileUploadHelper');
-const { uploadFile } = require('../../utils/fileUploader');
+const { mapSingleFileLocation } = require('../../middleware');
+const { uploadFile } = require('../../utils');
 
 router.post(
   '/',